Versioned CloudManager and NetworkManager collectors
This commit refactors collectors for CloudManager and Networkmanager,
so that the differences in collector methods due to our support for
two version profiles are accounted for in seperate subclasses of
manager's collector. We introduce one new class per API version profile
per manager. Behavior for the latest supported profile is kept in
Collector::CloudManager and Collector::NetworkManager, the profile
version-specific differences are covered in their subclasses. The
Inventory class now overrides `collector_class_for` method so that
the appropriate version of collector is constructed for CloudManager
and NetworkManager targets.

Signed-off-by: Manca Bizjak <manca.bizjak@xlab.si>